stable-patches: 4.4.172: Create based on -rc1
The f2fs conflicts come from a series of backports to fix some security
issues. I am not fighting through the conflicts, which come from the
fact that the f2fs maintainer now works for Google and has been merging
the f2fs-stable tree into kernel/common, which in turn gets merged into
kernel/common. I don't know of any 4.4 devices that ship with f2fs so
fighting through this merge is pointless and will result in more
conflicts down the road. We'll just wait for the f2fs-stable merge to
filter down to these trees in due time.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>