stable-patches: 4.4.187/4.9.187: Omit https://git.kernel.org/linus/56f3ce675103e3fb9e631cfb4131fc768bc23e9a
Most trees have [1] but not [2], which causes a compilation error
because of a warning:

../fs/f2fs/segment.c:3006:4: error: void function
'read_compacted_summaries' should not return a value [-Wreturn-type]

While this is a simple error to fix on the surface (just remove the
-EFAULT return value), it can defeats the purpose of that commit which
is to return the EFAULT up the stack, which won't happen without [2],
and it will cause another conflict in a later f2fs merge when [2] comes
down.

This patch will come back down in f2fs-stable [3], which is more
appropriate for it.

[1]: https://git.kernel.org/linus/c376fc0f3581c0b08099b0b7bfa448c30bc71c0c
[2]: https://git.kernel.org/linus/7735730d39d75e70476c1b01435b9b1f41637f0e
[3]: https://git.kernel.org/jaegeuk/f2fs-stable/c/e52e9e4cf21980bdd4224cd63c1136ded19071e3

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>